requirements.txt
fastapi
uvicorn
sqlalchemy
pytest
httpx
pydantic
jinja2
passlib[bcrypt]
python-jose[cryptography]
cryptography
python-multipart
bcrypt==3.2.2
google-genai
python-dotenv